refactor logger service to logger class (can be used outside of angular)
