+++ src/app/app/app-providers/fetch-config-providers.ts
@@ -6,12 +6,12 @@ import { layersConfig } from '@ansyn/menu-items/layers-manager/services/data-lay
 import { filtersConfig } from '@ansyn/menu-items/filters/services/filters.service';
 import { casesConfig } from '@ansyn/menu-items/cases/services/cases.service';
 import { LoginConfig } from '@ansyn/login';
-import { LoggerConfig } from '@ansyn/core/services/logger.service';
 import { enableProdMode } from '@angular/core';
 import { mapFacadeConfig } from '@ansyn/map-facade/models/map-facade.config';
 import { IdahoOverlaysSourceConfig } from './overlay-source-providers/idaho-source-provider';
 import { VisualizersConfig } from '@ansyn/core/tokens/visualizers-config.token';
 import { MultipleOverlaysSourceConfig } from './overlay-source-providers/multiple-source-provider';
+import { AnsynLogger } from '@ansyn/core/utils/ansyn-logger';
 
 export const getProviders = (conf): any[] => {
 	return [
@@ -64,16 +64,16 @@ export const getProviders = (conf): any[] => {
 		{
 			provide: MultipleOverlaysSourceConfig,
 			useValue: conf.multipleOverlaysSource
-		},
-		{
-			provide: LoggerConfig,
-			useValue: conf.loggerConfig
 		}
 	];
 };
 
 export const fetchConfigProviders = fetch('/assets/config/app.config.json')
 	.then(response => response.json())
+	.then(conf => {
+		AnsynLogger.activeLogger = new AnsynLogger(conf.loggerConfig.env);
+		return conf
+	})
 	.then(conf => {
 		if (conf.production) {
 			enableProdMode();
