+++ src/app/app/app.module.ts
@@ -6,10 +6,12 @@ import { AnsynModule } from './ansyn/ansyn.module';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { BrowserModule } from '@angular/platform-browser';
-import { GlobalErrorHandler } from '@ansyn/core/utils/global-error-handler';
+import { AnsynLogger } from '@ansyn/core/utils/ansyn-logger';
+import { GlobalErrorHandler } from './app-global-error-handler';
 
 export function MetaReducer(reducer) {
 	return function (state, action) {
+		AnsynLogger.activeLogger.info('action: ' + JSON.stringify(action));
 		return reducer(state, action);
 	};
 }
