+++ src/app/packages/core/core.module.ts
@@ -8,7 +8,6 @@ import { StoreModule } from '@ngrx/store';
 import { coreFeatureKey, CoreReducer } from './reducers/core.reducer';
 import { ToastComponent } from './components/toast/toast.component';
 import { ProjectionConverterService } from './services/projection-converter.service';
-import { LoggerService } from './services/logger.service';
 
 const coreComponents = [
 	AnsynCheckboxComponent,
@@ -24,8 +23,7 @@ const coreComponents = [
 	],
 	providers: [
 		GenericTypeResolverService,
-		ProjectionConverterService,
-		LoggerService
+		ProjectionConverterService
 	],
 	exports: coreComponents,
 	declarations: coreComponents
